Refactor containers for raw values
Split containers into left and right half in order to
speed up getting middle values

+++ src/types.hpp
@@ -1,9 +1,9 @@
 #pragma once
 
+#include <cstdint>
 #include <list>
 
-using RawValuesContainer = std::list<int>;
-
+using RawValuesContainer = std::list<uint16_t>;
 
 enum class ThresholdTrigger
 {
